Remove v3 files
