+++ components/brave_wallet_ui/page/screens/onboarding/restore-from-recovery-phrase/restore-from-recovery-phrase.stories.tsx
@@ -7,7 +7,7 @@ import * as React from 'react'
 
 import { WalletPageStory } from '../../../../stories/wrappers/wallet-page-story-wrapper'
 import { OnboardingRestoreFromExtension } from './restore-from-extension'
-import { OnboardingRestoreFromRecoveryPhrase } from './restore-from-recovery-phrase-v3'
+import { OnboardingRestoreFromRecoveryPhrase } from './restore-from-recovery-phrase'
 
 export const _OnboardingRestoreFromRecoveryPhrase = () => {
   return (
